+++ software/tracksight/live_data/app/read_task/log.py
@@ -1,0 +1,4 @@
+import threading
+def get_log_read_task() -> threading.Thread:
+    pass
+
