--- software/tracksight/live_data/app/write_task/log.py
@@ -1,4 +1,0 @@
-import threading
-def get_logging_task() -> threading.Thread:
-    pass
-
